Add coverage and test badges
In order to inform the users about the coverage and test passing of our repo, we are adding it using `coveralls.io`
and `excoveralls` package.

The badges will be added to the README so they can be seen easily.

Last, but not least, added the command `mix coveralls.github` into the CI so the results of the tests are sent to coveralls.